[web] Drop Agama/KebabMenu
Because it became a dummy wrapper after the changes introduced to the
Dropdown comonent by PatternFly.

The only thing that might make some sense to extract (from
Agama/WifiNetworkMenu) is the custom toggler if we end up using it in
more places.
