+++ web/src/assets/styles/app.scss
@@ -106,13 +106,11 @@ button.hidden-popover-button {
   word-break: keep-all;
 }
 
-.wifi-network-menu {
-  button.toggler {
-    padding-inline-start: 0;
+button.kebab-toggler {
+  padding-right: 0;
 
-    svg {
-      vertical-align: middle;
-    }
+  svg {
+    vertical-align: middle;
   }
 }
