+++ web/src/components/core/index.js
@@ -38,7 +38,6 @@ export { default as IssuesLink } from "./IssuesLink";
 export { default as IssuesPage } from "./IssuesPage";
 export { default as SectionSkeleton } from "./SectionSkeleton";
 export { default as LoadingEnvironment } from "./LoadingEnvironment";
-export { default as KebabMenu } from "./KebabMenu";
 export { default as LogsButton } from "./LogsButton";
 export { default as FileViewer } from "./FileViewer";
 export { default as RowActions } from "./RowActions";
